pyproject.toml
[project]
name = "terminal-games"
version = "0.1.0"
description = "Some simple terminal games"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "textual-dev>=1.7.0",
]
